ErrProp now parallelized
The error propagation code has been parallelized slightly. In FitObj the
loop for event reweighting has been split in two. The first loop just
performs reweighting, and is now parallel (same as the fit). The second
loop just fills the signal histogram(s), and is not parallel due to
using ROOT classes.

Few minor tweaks. Most notably XsecCalc now has a SetOutputFile function
and SaveOutput now only has one parameter for saving the toys in the fit
file.

xsllhCalc has a couple new command line options: one to set the
covariance matrix to use the mean of the toys in the calculation, and
one to save the toys in the output file.