ErrProp now can reweight flux histogram for normalization.
XsecCalc now has the capability to use the flux hist and flux parameters
to do the flux normalization.

FitObj has a ReweightFluxHist which takes the entire toy throw (or best
fit parameters) and picks out the flux parameters. It then reweights the
flux histogram, computes the integral, and returns it as a double.

The flux histogram for each signal is stored in the SigNorm struct, and
saved in the output.

The SigNorm structs also now contain histograms to show the distribution
of flux throws and number of target throws, and are saved in the output.

A slight change to FluxParameters was required to accomplish the
ReweightFluxHist function. The GetBinIndex() function became a public
function instead of private. Same thing with FitParameters for
consistancy.

New script to perform the POT reweighting of the official flux
histograms.

CMakeLists finally has an updated version number.