ErrProp saves more histograms.
The error propagation code now saves more histograms. It saves both the
concatenated signal histograms and the signal histograms separately
still by bin number.

It now has an extra, optional, parameter in the config for extra
histograms to save from the fit output file. It's a simple text file
where each line is the name of a histogram. If the string is empty, or
not present in the config file, it is ignored.

Finally it now has a function to plot the signals in 1D slices of their
2D binning, with actually correct units. This currently only works for
2D binning. It also currently sets the axis range to not plot the last
bin. This will probably be changed in the next commit as the plotting
script should take care of that.

BinManager was updated to add a function to access the underlying
bin_edge vectors to achieve the histogram plotting.