New config options for minimizer settings.
The fit now has support for changing a number of the minimizer settings
through the configuration file. All of these new settings are optional.

OptParser has a new struct, MinSettings, to store the new minimizer
options. XsecFitter has a new function which takes a MinSettings object
and stores the new options. In addition the constructor for XsecFitter
sets default values.

OptParser attempts to read a "min_settings" key in the JSON, and if
found reads any options it can find, using default values if one is
missing. If "min_settings" can not be found, the exception is caught and
the default values are loaded instead.

New config file commited with example. xsllhFit updated to include new
options.