ErrProp now has ability to throw from "prefit" covariance.
Added the necessary code for the error propagation to use a "prefit"
covariance matrix instead of the post fit.

This prefit covariance is built by stitching together the input
covariance matrices for the inputs, plus creating a statistical
covariance for the template parameters.

The template parameter "covariance" matrix is simply a diagonal matrix
using the statistical error for the analysis bins. This needs to be
squared, but it will be fixed next commit.

The error propagation code now checks whether to use the pre/postfit
covariance when initializing the toy thrower.

The prefit matrix if used is saved in the final output. xsllhCalcXsec
has a command line flag to enable using the prefit errors.

Needs more validation, but it's past time to leave.