Added options to enable/disable fitting systematic parameters
New options to enable/disable fitting the systematic parameters using
the config file. Set the "fit_par" option to true to include the
parameters, and false to not include them. If disabled, the parameters
are not included in the fit in any way, rather than fixing the
parameters.

xsllhFit and the error propagation code has been edited to include this
functionality.

Updated version in CMakeLists (finally). Update config with above
options.