Bug fix for detector cov; improved fit terminal output
A bug fix for unindented behavior in the detector covariance code. The
map that stores the binning files uses a std::string for the key, which
then is sorted alphabetically rather than numerically. This causes the
sorting to go 1, 10, 11, 12, 3, etc. in the case of numbers of 10 or
higher.

Slight tweak to the output of the fit parameters when applying PCA. The
index for how many parameters was off by one.

The fit init function now prints out the total number of parameters
along with the number of free and fixed parameters.

New output if the fit does not converge or if Hesse fails. An error
message is displayed along with the status code from the minimizer. If
either fails, the function returns immediately rather than continuing.