Fix bug when using stat fluctuations; Fix CMake bug
Fixed a bug when using statistical fluctuations and specifying an
external data set. The bin index was off by one because for a histogram
the first bin is at index 1, not 0.

Fixed CMake bug because when compiling with static libraries the order
of them matters when compiling. There is a more elegant fix for this,
but manually reording the targets will work for now.

When the fit does not converge, it now instead of exiting immediately
finishes the function to save all the output at the given iteration.
The XsecFitter::Fit() function now returns a bool for if the fit
converged or not. More error messages appear if the fit fails.