Added new decomposition routines to ToyThrower
Added a self-implemented version of both Cholesky and Incomplete
Cholesky decomposition. Both follow a slighly edited version of the
algorithm in Numerical Recipes in C 2nd Ed.

The incomplete Cholesky decomposition simply fills the L matrix with
zeroes in the same places the covariance matrix has zeroes (a fixed fill
algorithm). The method can also zero out entries in the covariance matrix
subject to a tolerance parameter.

New methods to set the covariance matrix without having to make a new
ToyThrower have been added.

New option in the ToyThrower constructor to control whether the ROOT
version of Cholesky decomposition is done automatically, or whether a
decomposition routine has to be called manually.