New errprop options; fixed calls to ToyThrower.
Fixed instances of using the ToyThrower since its constructor changed.

Errprop now has new configure options relating to how the decomposition
is done. Options to set type of cholesky decomposition and whether to
force the matrix positive definite.

The initalization of the ToyThrower in XsecCalc has changed to use the
mentioned options.