Save vectors of prefit pars; Truth xsec calc in Errprop.
The fit now saves TVectors of the prefit parameters, both the original
and decomposed values. If no parameters are decomposed, the two vectors
will be the same.

The truth cross-section is now calculated in the error propagation code.
Previously the true number of events were saved, rather than applying
the normalization to get a cross-section.

In addition, the true events are now reweighted using their nominal MC
weight instead of the post-fit parameter values.