Efficiency calculation updated.
Major update to the efficiency calculation in the error propagation
code. Well, kind of.

The efficiency is calculated between the selected and truth events
weighted using the post-fit parameter values (which is still the same as
before).

However the saved truth cross-section is the nominal true cross-section
using the original MC weights (no post-fit parameters). This gives a
reasonable looking chi-square for statistical fluctuations.

So in a way the efficiency calculation is the same, but now the correct
true cross-section is saved.

The error propagation code now reads and saves the vectors of the prefit
parameter values.