New command line options for xsllhFit.
Added some new command line options for xsllhFit which override the
options set in the JSON config.

The output file name, RNG seed, and number of threads can now be
specified using the command line. Values from the command line are used
instead of the corresponding JSON config values.

The command line values are still printed to the screen.