Updates to job submission script for job arrays.
Updated the job script and the submission script to submit a job array
instead of a single job.

A single job can still be used by only specifying one job when running
the submission script.

All jobs use the same configuration file, but the output file and RNG
seed change based on the job ID index.