Further updates to job scripts. New script for chisq calculation.
A couple more updates to the job submission scripts. The output files
now use the seed as part of the file name.

The script to setup the job file now has the ability to specify the seed
offset or use a random value.

New script to calculate the chisquare between two histograms,
specifically the selected and truth cross-sections from the error
propagation output.