Update to true xsec reweighting for systematic throws.
When doing systematic toy throws, the truth cross-section needs to be
reweighted using the systematic toy parameters from the fit.

The error propagation code now checks what type of fit was run (more
accurately, it checks if systematic toys were thrown) and then
reads the parameter vector for the throw.

The true cross-section uses the thrown parameter vector or nominal
values as appropriate.

Edits to how the input file is read and overrriden using CLI arguments.
This fixes an annoying bug were the an old file will be read and cause
a segfault if it is missing something.