Fit code now conditionally reads cov matrices
The main fit code (xsllhFit) now conditionally reads the covariance
matrices instead of always attempting to read them.

If the systematic parameter set is enabled in the config file, the
covariance file will be read. If disabled, then no attempt is made to
access the file.

This allows for testing the fit in parts. In addition slightly better
error handling was added if a covariance matrix file could not be
opened.