Adding T2KReWeight Code
Adding the T2KReWeight code used to create the cross-section spline
inputs.

The three main executables are included along with their versions of the
BinManager and XsecDial classes. The BinManager and XsecDials are
slightly different from the ones used in the fit code.

Eventually they will be edited to match the fit versions as closely as
possible.

The Makefile to compile the code is included. It can either replace the
default one with T2KReWeight or the specific lines can be copied.

Finally the script to run the whole process and examples of the dial
settings files are included.