Update to reweight code. Slight docs update.
Updated the reweighting code to change/fix how the dial values are set
when reading the errors as limits.

Previously, the upper and lower limit were used to create a range and
that was used to set the increment for determining dial values. This
works fine if the limits are symmetric. If the limits are not symmetric,
this can produce dial values above/below the specified limits.

The code was changed to determine the dial increment separately for
above and below the nominal value. This ensures the dial values will
stay within the specified limits.

Additionally a short doc page was added for the reweight code.
Eventually it will describe how to use it.