Added proper error return code to xsllhFit.
The main fit code, xsllhFit, now returns a non-zero error code if the
fit does not converge. This can be tested in a script to test if the fit
converged for later processing or scripts.

Slight update to documentation detailing minimum ROOT 5 version.

Commiting calc_chisq.cxx so Git stops telling me that it is modified.

Deleting unused xsec cov text file.