Updates to T2KReWeight code.
Important fix to the reweighting code if using numu flattrees instead of
OA files as the input to Highland.

When using numu flattrees, the association between the RooVtxIndex in
the default/truth tree and the proper vertex in the NRooTrackerVtx tree
is broken. Many events will not have a vertex found if using the
RooVtxIndex.

Instead all vertices must be looped through and checked if the
TruthVertexID's match between the vertex and the value from the
default/truth tree.

This has been added and through preprocessor definitions the original
code can still be used for OA files, which is slightly faster.

Also updated with an option to generate dummy spline inputs. Splines are
made but no weights are calculated (they are all set to 1.0). This gives
properly formatted splines for the fit for testing purposes.