ErrProp now conditionally reads covariance matrices.
The error propagation code now conditionally reads the covariance
matrices if they were used in the fit. Previously it always tried to
read them, causing failure if they were not used in the fit.

Now the whole fit and error propagation can be tested with none or some
of the systematic parameters.

Fixed the saving of extra histograms from the fit file where if the
histogram was not present the code just segfaulted. Now it checks for it
being a nullptr before writing to the file.

Slight output update for the T2KReWeight code. It now informs you which
method is being used to find the NRooTrackerVtx and if dummy splines are
being produced.