Adding a function to cap parameter weights.
Added a function to AnaFitParameters to set variables for capping event
weights while reweighting.

The xsecparameters reweight function has code to cap the weights. Flux
and detector parameters currently ignore the weight cap.

This is not enabled anywhere in the fit or error propagation code as of
this commit. Enabling would require adding the SetWeightCap() function
to the executables.

Small update to the fit job script to check if the fit successfully
converged or not before running the error propagation.

Finally updating the CMake project verison.