Added capability to ErrProp to read from data file.
Added the functionality for the error propagation code to read events
from the data file used in the fit.

This is for fake data studies to easily plot the true fake data
cross-section in the same binning, etc. as the fit.

This can be enabled through a configure option.

Also added the ability to save the MC truth cross-section.

Changed the function to save the signal histograms in the physics
binning to take a vector histograms as an option rather than hardcoding
it to only use the selected event histograms.