First pass at master plotting code.
As it says, the first pass at writing a master plotting code to produce
all the plots at once.

Currently it reads a small JSON config and creates some of the
cross-section plots.

A new set of utility classes have been written to control the global and
specific histogram styles.

CMake files have been updated to compile these new files.

Going to try a new method to make the plots...