Further updates to xsllhPlot
Added a number of new features to the master plotting code.

Can now set axis ranges and set either to logrithmic scale.

Added support for using a default file to read histograms from. Can be
specified at the beginning of the configure file or at the command line.
The code then reads from this file unless one is specified for the plot,
which overrides the default behavior.