Updates to plotting code. Minor change to FitParameters.
Mostly updates to the plotting code.

Changed the behavior for the FitParameters when an event does not fall
in an analysis bin. Previously the event was given a weight of zero,
effectively removing it entirely. This causes issues when calculating
the chi-square because it is still counted in the calculation.

Now the weight is simply unchanged by the template parameters if it does
not fall in an analysis bin.