Fit now saves globalcc and new options for rng start point.
The fit code now saves the global correlation coefficient for each
parameter in the output as a TVectorT.

New/redesigned options for having the fit not start at the best fit
point. This does not change the prior values, only the initial value of
each parameter for the minimization. This is an important distinction
due to how the covariance penalty terms use the prior in their
calculation.

Two new options: a "rng_template" option and "rng_start" option for each
of the systematic parameters. This randomizes the initial value passed
to Minuit by adding/subtracting a fractional value from a Gaussian
distribution with 10% width.

The start point values are also saved in the fit output as
vec_prefit_start.