New function to perform parameter scans.
New function in XsecFit to perform parameter scans and save the output.
The config file now has two new (optional) settings: par_scan_list, and
par_scan_steps.

The parameter scan is performed by moving the selected parameter and
calculating the likelihood function, holding all other parameters fixed.

'par_scan_list' is a list of parameter numbers to scan after the
minimization is finished. 'par_scan_steps' is the number of steps to use
in the scan. By default it is 20 steps.

The scans are saved as TGraph's in the final fit output.