Modifications to saving events in fit output.
Changed the variables when saving the fit events. Removed weightNom and
made both weight and weightMC to include the POT weighting.

Also added a SignalType variable to the output.

New option in the configure file to enable/disable saving the events in
the fit output file. Corresponding code added to XsecFitter and
OptParser.