Reorganized saving event tree in XsecFitter
Realized the final reco event histograms were also being saved in the
function that saves the TTree of the events.

Reorganized the function so it only saves the TTree, and thus correctly
does what was intended when enabled/disabled.

Combined saving the final reco histograms with the function that would
save them for any given iteration.

Renamed the functions to better describe what they do.