Updated Progress bar to check if stdout is tty.
Updated and refactored ProgressBar to include a check if stdout is a TTY
or being redirected to something else.

If stdout points to TTY, then perform the normal updating bar. If not,
   then print plain text numbers and no '\r' characters.

This check is controlled by a CMAKE flag which is on by default since
the 'isatty()' function is non-standard C (but standard POSIX).