Added more histograms to errprop output.
New histograms containting the relative error on the post-fit
cross-section have been added to the output of the error propagation.

Both the linear bin number and pmu-cosmu binnings have histograms with
the relative error. This is for easy plotting and comparison of the
post-fit errors.

Also updated the range to include all bins, rather than hiding the last
bin. This step (if desired) will be moved to the plotting code.