New likelihood functions and implementation
New likelihood functions have been added to include the effect of the MC
statistical error -- the Barlow-Beeston method and the effective
likelihood from Tianlu et. al.

A new header defining all the likelihood functions as functor classes
has been included. This way a likelihood function can be instantiated
during initalization instead of an if/else branch for every calculation
of the likelihood.

This also gives better organization for adding or modifying the
likelihood functions.

A new config option has been added to OptParser to change the
likelihood. The code by default uses the normal Poisson likelihood, and
if an invalid likelihood option is supplied, it will also default to the
normal Poisson one.

Updates to calc_chisq.cxx to produce nice candle plots of the cross
section distribution and the errors for each bin.