Updates to calc_chisq, change to build setup script.
Updated calc_chisq.cxx to include the number of failed fits as defined
by some chisq cutoff value. Improved visual look of plots. Adds a chisq
distribution based on the fit to toys.

Slightly edited the build script which is auto-generated by CMake to fix
a bizarre bug seen on other platforms. Somehow when sourcing ROOT after
editing the paths to include the fit, ROOT incorrectly changes the
fitter paths. Sourcing ROOT first fixes this problem.

It shouldn't matter at all when/where ROOT is sourced, but... this works and
we're just going to roll with it.