Add the NVIDIA hemi template library
The hemi template library is distributed by NVIDIA with a free
distribution liscence (included in this commit).  This library
provides the infrastructure to manage memory between the CPU and
GPU, and to use the same code to write kernels on both.  It is written
in the CUDA subset of C++, and can be compiled with g++, clang, or
nvcc.  If nvcc is used, the code can run on a CUDA gpu.