Initialize GPU before first use of Propagator::reweightSampleEvents
There are two important changes in this commit.  First, the GPU cache
is filled right after the MC might have been copied to the data set for
an Asimov sample, but before the first call to reweightSampleEvents.  This
guarrantees that the cache is used for all reweighting.  Second, the
initial result value is explicitly set in buildGPUCache so that it doesn't
depend on the use of PhysicsEvent::setTreeWeight().