PhysicsEvent::setTreeWeight does not touch GPU cache
The CachedWeights initial value must not be explicitly set, and there
is no implicit synchronization with the physics event tree weight.
There is an optional sanity check in PhysicsEvent::getEventWeight, that
should probably be disabled for productions because it adds a bit of time
to the calculation (and is called *a* *lot*).