Fully remove update of GPU cache from PhysicsEvent::setTreeWeight
This deletes the code lines (not just ifdefs) to change the GPU cache
in PhysicsEvent::setTreeWeight.  The original code had the problem that
it might set the treeweight for the wrong cache entry, so it was buggy.

Expand the comment that the debugging code in PhysicsEvent::getEventWeight
is extremely expensive and should not usually be used.