Move GPUInterpCachedWeights into separate library
This is the beginning of refactoring the GPU cache so that it can
handle all of the GPU calculations we will (probably) want.  Possible
additional calculations are: splines with input slopes, splines with
non-uniformly spaced knots, accumulation of histogram entries,
assigning events to histogram bins, calculation of oscillation
parameters, etc (to name a few).