Make the usage of Cached::Weights more thread safe
The Cached::Weights management of the cached values was depending
on copying the results from the GPU as soon as they were updated
and added a fixed copy time to applying the kernels.  Because the
weights only need to be copied if the host is accessing the event
weight, a bit more complexity has been added to copy the weights
"as needed", but in a thread safe way.

There is minimal (if any) change in the processing speed.