Use doubles for internal spline calculations
The inputs to the splines can be either "float" or "double" depending
on the value of WEIGHT_BUFFER_FLOAT (currently set to "double").  It
might be set to "float" to save memory, but the internal calculations
will always be promoted to a double so that less numerical precision is
lost.