Allow a DialSet subtype to be set in the dial configuration
This adds a yaml field to the DialSet definition to allow a subtype
to be specified.  The first use is to determine the type of spline that
will be used (e.g. a monotonic (compact) spline, or a natural spline), but
It will likely be more important as we move away from depending on
TSpline3 by default.