Use new spline calculation functions in place of TSpline3
This uses CalculateUniformSpline() and CalculateGeneralSpline() in place
of TSpline3 in the SplineDial.  The spline result have been validated against
TSpline3, and remain inside the machine accuracy.  This has the consequences that
the CPU and GPU calculation can use the same source code to calculate the spline,
splines types not provided by TSpline3 (e.g. monotonic) can be implemented
generally, and in "out-of-job" benchmarking the new routines are faster than
TSpline3 (however, negligible impact is seen when used in GUNDAM).