Fix to compile with refactored NormDial
The normalization dial was refactored, and changed name.  This fixes
some places that the refactoring missed, and that no longer compile.

NOTE: This continues to use a downcast, but it seems like there are
beginnings of a new internal RTTI interface to simplify dial type
inference.  When that interface is stabilized, this file should be
updated.