Add WITH_GENERIC_SPLINES cmake option (def: off)
This removes the hardcoded definition for USE_TSPLINE3_EVAL from
SplineDial.h and makes it an optional flag in the CMakeLists file
(it's defined unless WITH_GENERIC_SPLINES is on, or WITH_CACHE_MANAGER
is on).  WITH_GENERIC_SPLINES=off is ignored with WITH_CACHE_MANAGER
is one.

This includes some changes for SplineDial.  Mostly it is making sure
the ifdef logic is clean, but also adds an explicit option to request
TSpline3 when generic splines are used.