BUG FIX: Check validity of _parameterNamesList_
The _parameterNamesList_ pointer normally points to the TObjArray
named by parameterNameTObjArray holding the index names of the
constraint covariance matrix.  The FitParameter name was being filled
assuming that _parameterNamesList_ pointed to a valid TObjArray

Check that _parameterNamesList_ is a valid TObjArray pointer.  If it
is, then get the FitParameter name and configuration from the
TObjArray (the old logic).  If _parameterNamesList_ is a nullptr, then
get the name and configuration from the parameter definition
configuration list (i.e. YAML).