Add warning for parameter not in covariance matrix file
Print a warning when a fitting parameter is defined, but does not
exist in the covariance matrix file.  An example of where this might
happen is when an unconstrained normalization parameter is defined.

The missing parameter is actually triggered by a missing
parameterNameTObjArray field in the YAML, and a warning is generated
since this field should usually (but is not required to) exist.