Expand pathnames from the config files
The yaml was only accepting fully resolved absolute and relative
paths.  This adds the expansion of environment and shell globs using
GenericToolbox::expandEnvironmentVariables.  A warning message is
printed when a file name is expanded.

Environment variables are used by the new validation code because the
run location relative to the configuration and input files may not be
known until runtime.