Change fit tests to PoissonLLH and 1E-6 tolerance
The fitting tests were using the BarlowLLH_BANFF_OA2021 joint
probability (good for physics) and now use PoissonLLH.  This is done
so that the tests are mathematically simpler and focussed on the rest
of the infrastructure.

The tolerance for the test was increased since the numerical stability
seems to be about 5E-7.