Use MinimizerBase inside FitterEngine
This changes the time of the _minimizer_ field in FitterEngine to use
MinimizerBase instead of MinimizerInterface.  That forced makeing the the
minimizer parameter scan into a virtual in MinimizerBase and
implementing it in the MinimizerInterface.  It also meant turning
isFitConverted into a MinimizerBase virtual method and enabling the
scan into MinimizerBase.