Add MCMCInterface to run an MCMC chain
This adapts Shilin's original GUNDAM-MCMC implementation to be
derived from MinimizerBase, and use the LikelihoodInterface class
to calculate the likelihood.  It can be called by the
FitterEngine as an instantiation of the MinimizerBase class.
The MCMCInterface doesn't produce a covariance.  Instead it
produces a point cloud that describes the Bayesian posterior
probability distribution.