Add more testing levels to gundam-tests.sh
This adds the regular and extended testing level to the tests.
The fast tests are run during CI.  Regular tests should be run
before pushes and pull-requests.  The extended testing can
take a few minutes, so it's run as part of development
validation.  The slow tests can take a very long time so they
are intended as part of release validation.