Add configuration override from command line
Values previously set in a configuration file can be overridden
from the command line using the '--override key=value' option.
The override is done immediately after the configuration is
read, and will be saved in the output file as part of the unfolded
configuration.  The key must preexist in the input YAML/JSON or
an error will be thrown (this is intentional).