Temporary bug fix: Add missing include files
The USE_NEW_DIALS ifdefs were incorrectly done, so the definitions
were ignored in FitParameterSet.cpp and SplineDial.cpp.  That broke
after the definition was correctly defined for the build.  The
temporary solution is reproduce the bug by explicitly includeing the
old dial definitions.  That will need to be fixed, and should
naturally happen when the correct deprecation flags are set.