Deprecate methods in new dial implementation
Rather than simply remove some of the methods that are not
applicable to the new dial implementation, and then have
compilations fail "oddly", this now marks them as deprecated so
that the compiler has a decent chance to diagnose the problem and
provide a good error message.  This also throws an error if one of
the method is used during runtime.