Add DialBase factories to limit code duplication
There were a few separate places where the dial type to dial class
mapping was being done, and there were starting to be small
inconsistencies between the different bits of code.  This isolates
all of the DialBase object creation into a single set of factories
that can be used everyplace.