Update the Calculate*Spline documentation
The function header comments have been updated to be more accurate
(not a copy and paste), and to call out the differences in the
function interfaces.  The interfaces are not changed (yet) since
that needs to be tested on the GPU (and I don't have the development
environment working for a quick fix).