Make sure gRandom is properly initialized
The gRandom "variable" was being initialzed by default.  That
usually means it's assigned to TRandom3, with a seed of 1.  The
seed was then being set off of the time which caused problems when
starting a lot of parallel jobs.  Changed to explicitly initialize
gRandom to TRandom3(0) so we get a well known generator that is
initialized from the UUID.