Record the mirroring used for a parameter in FitParameter
The mirroring is coded as part of a dial, but is also closely related
to the FitParameter.  This adds an accessor to track which mirroring
is being applied to a parameter.  Since mirrored parameters (when they
exist) mostly apply to a single dial, there is a warning when a
parameter uses multiple mirror points.