Separate usage and storage of the accepted step
Add a separate storage object for the accepted step so that the
internal fAccepted variable is not exposed to users through the
output tree.  The copy that is saved can be cleared by the user
and save space in the output (e.g. if the user is processing the
accepted step before saving and does not need the raw value).