Record parameter physical limits
This adds getters and setters to FitParameter for the parameter
physical limits.  This is used to record the range over which a
parameter has physical meaning (separately from where the parameter
has a mathmatical definition).  The limits can be set from YAML using
the physicalLimits.minValue and physicalLimits.maxValue entries.