Make compact monotonic extrapolation symmetric
The original implementation was a stripped down version of the
Fritsch-Carlson criteria, but it was't symmetric in how it
handled the end points, so the extrapolation of a symettric
function was asymmetric.  This now does the full calculation, and
has almost identical benchmarks (no measurable difference).

Test are added for the CalculateMonotonicSpline routine, and the
test for CalculateCompactSpline was updated so it's the same.