Don't disable error checking
Somebody disabled an error trap (probably while implementing
something) and we could run fits clearly wrong results.  This
traps when the config can't be valid.  In this case, the trap
was for a dial type that had a missing implementation.  I found
it by inspection, so it is probably harmless.