Resolve 522: Back port tests from main to LTS (#523)
Resolve 522

* Backport test : 900NormalizationMCMCCheck.C

This test is updated to take into account the MCMC effective sample
size when checking the expected values.  The effective decorrellation
length is a guess, but based on looking at the input sample.

* Backport test : 900HorrifyingMCMCCheck.C

This test is updated to take into account the MCMC effective sample
size when checking the expected values.  The effective decorrellation
length is a guess, but based on looking at the input chain.

* Update the change log