Add thread count and a direct calculation flag
Hold a couple of more command line oriented values that are needed by
the internals.  1) A flag to control if the CPU calculation is forced.
This can be used to run the GPU and CPU in parallel. 2) A separate
parameter for the total number of threads.  This can be used to limit
the number of threads used by utility libraries (for example, HEMI).