Fix incorrect filling of histograms
The logic was wrong when the histograms were filled before the
internal structures were finalized. This now makes sure that the
histogram gets the right values (no more NANs in the histogram).
There remains a question as to why the NAN values didn't cause the fit
to stop.  This also handles (optionally) running the GPU and CPU
calculation in parallel and makes sure they are getting the same
result.