Fix the event weight and histogram binning caches
The cache code was returning spurious and partial results for the
CPU and GPU depending on when it was called.  This fixes the logic
so the cache returns a valid result.  It also adds the possibility
to back the CPU calculation with the GPU.  This has the advantage of
the GPU code always doing a full calculation and having much simpler
logic so it is quite reliable as a cross check for catching cache
mistakes.