Add direct query for parameter masking
In the dials, the "set" for the dial is determined based on the
parameter sets contained in the input buffer.  If the parameter
set for a particular parameter is masked, the dial is masked.  Add
a Parameter query to check for matching so we don't have to copy the
code snippet.