Check if a parameter is masked and flag
The CacheManager can't change the set of dials used during a
calculation, so just *FLAG* that the caller has changed the dials to
be used.  This probably means that some of the bug checking code is
out of place and is being used after the likelihood is should be frozen
in place.