Fix: Don't change the Propagator
Make sure that the likelihood doesn't change the event weighting
function.  That is handled in the data set manager.  Elsewhere, the
propagator needs to be treated as "immutable" (meaning once it's
initialized by the data set manager, the rest of the libraries needs
to Leave. It. Alone.).  Any other path leads to madness.