Update building of 100NormalizationTree.root to match main
There was a difference found between the validation results in the
main branch and the LTS branch.  It was traced to the way the inputs
are generated in main.  Since the way in main is slightly better, and
it is good to keep the same test results in both branches, the main
method is backported to LTS.  This only changes the testing code.