Disallow parameter bounds with eigendecomposition
This flags an error (convertable to a warning) if a fit is using the
undefined behavior for a parameter set with bounded parameters and
using eigendecomposition.

The error can be converted into a warning by setting
"allowEigenDecompWithBounds" for a parameter set.  Out of bounds
parameters will still cause an error, but the fit will proceed as
long as the best fit, and the path to the best fit stay inside the
allowed boundaries.