Upgrade ParameterSet covariance throws to be safe
The parameter set throws were sloppy about setting parameter values
out of bounds.  This cleans up the code so that it is pedantic about
the bounds, and adds the ability to require throws to also be inside
the physical bounds for the parameter.  Warning messages are added
if a non-standard parameter thrower is used.