Add a way to print backtraces
This puts the backtrace onto a stream so so that we can log where
error messages are coming from.  It's used as a stream modifier, so
"LogInfo << GundamUtils:Backtrace;"
or "std::cout << GundamUtils::Backtrace;"
will work.  The backtrace is only available on linux for the moment,
and elsewhere it will output "Backtrace is not available".