Allow invalid parameter values to be forced
This adds an optional parameter to Parameter::setParameterValue to
allow an invalid parameter value to be set.  If the optional parameter
is not provided, or is set to false, then an invalid value will
terminate with EXIT_FAILURE.  If the optional parameter is true, then
invalid parameters will still produce error messages, but will
continue with an alert that an invalid parameter has been set.