Add validity check that GPU was run
Parameters can be masked when the status plots are built, but
since removing parameters can change the overall normalization,
it changes the likelihood definition, and the GPU cannot be used
without full reinitialization.  Since only a few calls have
masked parameters, skip the GPU in those cases and use the CPU
instead.