Check the likelihood definition was not changed
Some of the status dumps are masking parameter sets to help
understand the effect of the sample sizes.  Unfortunately, some
response functions change the overall normalization as the parameters
are enabled and disabled (i.e. the response can be a constant with
the nominal response weight not being 1.0).  The GPU cannot handle
that without reinitialization, so disable for those calls.  This has
negligible effect on the running time.