Expose the address for the dial update request
The access to the isDialUpdateRequested() is a significant
bottleneck due memory L1/L2/L3 cache misses.  This uses a "dirty
trick" to let the boolean be checked in a single memory access
instead of the current three.  It also limits the memory set that
needs to be in the cache.  Very preliminary testing shows a speed
increase of more than two.